Send data as its encoded row-by-row, not once at the end
This is how COPY is meant to be used, but we haven't actually done that
in the library, mostly for historic reasons. This is now the new default,
which can be disabled by passing the `send_at_once: true` keyword
argument to copy_from_client.

Note that this doesn't measurably affect performance in limited tests,
but does reduce peak memory usage when copying in a large number of rows.